New basic container tech test including podman
