+++ stimela/cargo/cab/ddfacet/parameters.json
@@ -1,7 +1,7 @@
 {
     "task": "ddfacet", 
     "base": "stimela/ddfacet", 
-    "tag": "1.3.0", 
+    "tag": "1.3.1", 
     "description": "DDFacet: a facet-based imager", 
     "prefix": "--", 
     "binary": "DDF.py",
